requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
greenlet==3.0.3
idna==3.7
pyTelegramBotAPI==4.18.1
python-dotenv==1.0.1
requests==2.32.2
SQLAlchemy==2.0.30
typing_extensions==4.12.0
urllib3==2.2.1
